test(Lotto): 로또 테스트 처리
패키지 위치는 model로 이동
- 로또 숫자 구성은 총 6자리
- 로또 숫자는 중복수가 있으면 안됨.
- 로또 숫자는 1~45 사이의 수로 구성
